bios/iccsense: the mohm field seems to be a little bigger

+++ drm/nouveau/include/nvkm/subdev/bios/iccsense.h
@@ -1,7 +1,7 @@
 #ifndef __NVBIOS_ICCSENSE_H__
 #define __NVBIOS_ICCSENSE_H__
 struct pwr_rail_resistor_t {
-	u8 mohm;
+	u16 mohm;
 	bool enabled;
 };
